agentic-data-analyst/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-community==0.3.14
langchain-core==0.3.30
langgraph==0.2.62
pandas==2.2.3
numpy==2.2.1
matplotlib==3.10.0
plotly==5.24.1
kaleido==0.2.1
chromadb==0.6.3
python-dotenv==1.0.1
python-multipart==0.0.20
